services/conhecimento-backend/pyproject.toml
[project]
name = "conhecimento-backend"
version = "0.1.0"
description = "Conhecimento RAG - Base de Conhecimento com pgvector e Gemini"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "greenlet>=3.0.0",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.9",
    "pymupdf>=1.24.0",
    "python-docx>=1.1.0",
    "openpyxl>=3.1.0",
    "httpx>=0.27.0",
    "pgvector>=0.3.0",
    "flashrank>=0.2.0",
    "pytesseract>=0.3.10",
    "pdf2image>=1.16.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.5.0",
]

[tool.setuptools]
packages = ["app"]

[tool.ruff]
target-version = "py311"
line-length = 100
